+++ report-service/src/main/java/com/example/reportservice/entity/outbox/OutBox.java
@@ -1,10 +1,7 @@
-package com.example.reportservice.entity.message;
+package com.example.reportservice.entity.outbox;
 
 import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
+import lombok.*;
 
 import java.time.LocalDateTime;
 
@@ -13,6 +10,7 @@ import java.time.LocalDateTime;
 @NoArgsConstructor(access = lombok.AccessLevel.PROTECTED)
 @AllArgsConstructor
 @Builder
+@ToString
 public class OutBox {
 
     @Id
