+++ report-service/src/main/java/com/example/reportservice/repository/outbox/OutBoxRepository.java
@@ -1,6 +1,6 @@
 package com.example.reportservice.repository.outbox;
 
-import com.example.reportservice.entity.message.OutBox;
+import com.example.reportservice.entity.outbox.OutBox;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface OutBoxRepository extends JpaRepository<OutBox, Long> {
